requirements.txt
requests
beautifulsoup4
twilio
google-cloud-vision
opencv-python
flask
gunicorn